refactor: description 상수 파일 중복 제거 및 통합
