--- src/shared/constant/stories-view.ts
@@ -1,7 +1,0 @@
-export const START_BEFORE =
-  "한국장기조직기증원 장기구득 코디네이터가 함께 한 영원히 기억되는 기증자의 숭고한 나눔의 순간과 아름다운 이야기를 적는 공간입니다.";
-
-export const CHECK_ITEMS = [
-  "KODA 장기구득 코디네이터들의 공간입니다.",
-  "개인정보 노출의 우려가 있으니 게시글 작성 시 개인정보 등록은 자제하여 주시기 바랍니다.",
-];
